# apps/api/internal/api/openapi.json
{
    "openapi": "3.0.3",
    "info": {
        "title": "PCA API",
        "version": "1.0"
    },
    "paths": {
        "/auth/login/wca/": {
            "post": {
                "operationId": "auth_login_wca_create",
                "description": "Login with WCA oauth authorization code",
                "summary": "Login with WCA oauth authorization code",
                "tags": [
                    "auth"
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/WCALogin"
                            },
                            "examples": {
                                "WCALoginRequestExample": {
                                    "value": {
                                        "code": "OTQ5HFpRcpwBJPxGZgwc0Dc5LpnTUXxVVNHxe2QDdEl",
                                        "callback_url": "http://localhost:8080/wca-callback"
                                    },
                                    "summary": "WCA login request example"
                                }
                            }
                        },
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "$ref": "#/components/schemas/WCALogin"
                            }
                        },
                        "multipart/form-data": {
                            "schema": {
                                "$ref": "#/components/schemas/WCALogin"
                            }
                        }
                    },
                    "required": true
                },
                "security": [
                    {
                        "tokenAuth": []
                    },
                    {
                        "cookieAuth": []
                    },
                    {}
                ],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/WCALogin"
                                },
                                "examples": {
                                    "WCALoginResponseExample": {
                                        "value": {
                                            "key": "cjcsu60bc8fi4hw9s4lbhgnbtc4ls1xlyga99qe0"
                                        },
                                        "summary": "WCA login response example"
                                    }
                                }
                            }
                        },
                        "description": ""
                    }
                }
            }
        },
        "/auth/logout/": {
            "post": {
                "operationId": "auth_logout_create",
                "description": "Signs the current user out and deletes their API token.\n\nAccepts/Returns nothing.",
                "summary": "Signs the current user out and deletes their API token.",
                "tags": [
                    "auth"
                ],
                "security": [
                    {
                        "tokenAuth": []
                    },
                    {
                        "cookieAuth": []
                    },
                    {}
                ],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/RestAuthDetail"
                                }
                            }
                        },
                        "description": ""
                    }
                }
            }
        },
        "/events/": {
            "get": {
                "operationId": "events_list",
                "description": "List of official WCA events",
                "summary": "List of official WCA events",
                "tags": [
                    "events"
                ],
                "security": [
                    {
                        "tokenAuth": []
                    },
                    {
                        "cookieAuth": []
                    },
                    {}
                ],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Event"
                                    }
                                },
                                "examples": {
                                    "EventListExample": {
                                        "value": [
                                            [
                                                {
                                                    "id": "333",
                                                    "name": "3x3x3 Cube",
                                                    "rank": "10",
                                                    "format": "time",
                                                    "cell_name": "3x3x3 Cube"
                                                }
                                            ]
                                        ],
                                        "summary": "Event list example"
                                    }
                                }
                            }
                        },
                        "description": ""
                    }
                }
            }
        },
        "/news/": {
            "get": {
                "operationId": "news_list",
                "description": "News Feed from Facebook Page",
                "summary": "News Feed from Facebook Page",
                "tags": [
                    "news"
                ],
                "security": [
                    {
                        "tokenAuth": []
                    },
                    {
                        "cookieAuth": []
                    },
                    {}
                ],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/News"
                                    }
                                },
                                "examples": {
                                    "NewsListExample": {
                                        "value": [
                                            [
                                                {
                                                    "from_name": "Philippine Cubers Association",
                                                    "message": "Hello World!",
                                                    "image": null,
                                                    "permalink": "https://www.facebook.com/182835918433029/posts/3709202639129655/",
                                                    "created_at": "2021-04-24T02:29:17.921Z"
                                                }
                                            ]
                                        ],
                                        "summary": "News list example"
                                    }
                                }
                            }
                        },
                        "description": ""
                    }
                }
            }
        },
        "/persons/{wca_id}/": {
            "get": {
                "operationId": "persons_retrieve",
                "description": "Retrieve WCA profile and statistics",
                "summary": "Retrieve WCA profile and statistics",
                "parameters": [
                    {
                        "in": "path",
                        "name": "wca_id",
                        "schema": {
                            "type": "string"
                        },
                        "required": true
                    }
                ],
                "tags": [
                    "persons"
                ],
                "security": [
                    {
                        "tokenAuth": []
                    },
                    {
                        "cookieAuth": []
                    },
                    {}
                ],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Person"
                                },
                                "examples": {
                                    "PersonRetrieveExample": {
                                        "value": {
                                            "id": "2021DELA01",
                                            "name": "Juan dela Cruz",
                                            "country": "Philippines",
                                            "gender": "Male",
                                            "avatar": {
                                                "url": "https://www.worldcubeassociation.org/assets/missing_avatar_thumb-f0ea801c804765a22892b57636af829edbef25260a65d90aaffbd7873bde74fc.png",
                                                "thumb_url": "https://www.worldcubeassociation.org/assets/missing_avatar_thumb-f0ea801c804765a22892b57636af829edbef25260a65d90aaffbd7873bde74fc.png",
                                                "is_default": true
                                            },
                                            "career": {
                                                "competition_count": 12,
                                                "solve_count": 60,
                                                "personal_records": {
                                                    "333": {
                                                        "single": {
                                                            "best": "12.32",
                                                            "world_rank": 14912,
                                                            "continent_rank": 4105,
                                                            "country_rank": 341
                                                        },
                                                        "average": {
                                                            "best": "14.02",
                                                            "world_rank": 15318,
                                                            "continent_rank": 5291,
                                                            "country_rank": 421
                                                        }
                                                    }
                                                },
                                                "records": {
                                                    "national": 3,
                                                    "continental": 0,
                                                    "world": 0,
                                                    "total": 3
                                                },
                                                "medals": {
                                                    "gold": 1,
                                                    "silver": 2,
                                                    "bronze": 0,
                                                    "total": 3
                                                }
                                            }
                                        },
                                        "summary": "Person retrieve example"
                                    }
                                }
                            }
                        },
                        "description": ""
                    }
                }
            }
        },
...
        "/regions/": {
            "get": {
                "operationId": "regions_list",
                "description": "List regions in the Philippines",
                "summary": "List regions in the Philippines",
                "tags": [
                    "locations"
                ],
                "security": [
                    {
                        "tokenAuth": []
                    },
                    {
                        "cookieAuth": []
                    },
                    {}
                ],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Region"
                                    }
                                },
                                "examples": {
                                    "RegionListExample": {
                                        "value": [
                                            [
                                                {
                                                    "id": "NCR",
                                                    "name": "NCR (Luzon - Metro Manila)"
                                                }
                                            ]
                                        ],
                                        "summary": "Region list example"
                                    }
                                }
                            }
                        },
                        "description": ""
                    }
                }
            }
        },
        "/statistics/growth/active-competitors/": {
            "get": {
                "operationId": "statistics_growth_active_competitors_retrieve",
                "description": "Prepared annual regional growth series.",
                "tags": [
                    "statistics"
                ],
                "security": [
                    {
                        "tokenAuth": []
                    },
                    {
                        "cookieAuth": []
                    },
                    {}
                ],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/GrowthMetricResponse"
                                }
                            }
                        },
                        "description": ""
                    }
                }
            }
        },
        "/statistics/growth/attendances/": {
            "get": {
                "operationId": "statistics_growth_attendances_retrieve",
                "description": "Prepared annual regional growth series.",
                "tags": [
                    "statistics"
                ],
                "security": [
                    {
                        "tokenAuth": []
                    },
                    {
                        "cookieAuth": []
                    },
                    {}
                ],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/GrowthMetricResponse"
                                }
                            }
                        },
                        "description": ""
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Event": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "string",
                        "maxLength": 6
                    },
                    "name": {
                        "type": "string",
                        "maxLength": 54
                    },
                    "rank": {
                        "type": "integer"
                    },
                    "format": {
                        "type": "string",
                        "maxLength": 10
                    },
                    "cell_name": {
                        "type": "string",
                        "maxLength": 45
                    }
                },
                "required": [
                    "cell_name",
                    "format",
                    "id",
                    "name",
                    "rank"
                ]
            },
...
            "Region": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "string"
                    },
                    "name": {
                        "type": "string"
                    }
                },
                "required": [
                    "id",
                    "name"
                ]
            },
...
            "RestAuthDetail": {
                "type": "object",
                "properties": {
                    "detail": {
                        "type": "string",
                        "readOnly": true
                    }
                },
                "required": [
                    "detail"
                ]
            },
...
            "WCALogin": {
                "type": "object",
                "properties": {
                    "code": {
                        "type": "string"
                    },
                    "callback_url": {
                        "type": "string",
                        "nullable": true
                    }
                },
                "required": [
                    "code"
                ]
            },
...
        },
        "securitySchemes": {
            "cookieAuth": {
                "type": "apiKey",
                "in": "cookie",
                "name": "sessionid"
            },
            "tokenAuth": {
                "type": "apiKey",
                "in": "header",
                "name": "Authorization",
                "description": "Token-based authentication with required prefix \"Token\""
            }
        }
    }
}